backend/requirements.txt
Flask==2.0.2
Flask-MySQLdb==0.2.0
flask_cors
python-decouple==3.5
requests==2.26.0
PyJWT
ecdsa
base58